==> 1. NBA Skills Improvement Project Part 1 - Data Scraping.ipynb <==
## NBA Skills Improvement Project Part 1 - Data Scraping

### Imports

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

import requests
import time
import json

from bs4 import BeautifulSoup

pd.set_option('display.max_columns', None)

years = list(range(2014,2024))

### Scraping from Basketball Reference

shooting_url = 'https://www.basketball-reference.com/leagues/NBA_{}_shooting.html'
defense_url = 'https://www.basketball-reference.com/leagues/NBA_{}_advanced.html'

for year in years:
    url = shooting_url.format(year)
    data = requests.get(url)
    
    with open("C:/Users/kevin/Jupyter Notebook Projects/shooting stats/{}.html".format(year), "w+", encoding='utf-8') as f:
        f.write(data.text)
        
for year in years:
    url = defense_url.format(year)
    data = requests.get(url)
    
    with open("C:/Users/kevin/Jupyter Notebook Projects/defensive stats/{}.html".format(year), "w+", encoding='utf-8') as f:
        f.write(data.text)

Testing with just 2023 to see how the data looks.

with open("C:/Users/kevin/Jupyter Notebook Projects/shooting stats/2023.html", encoding='utf-8') as f:
    page = f.read()
    
soup = BeautifulSoup(page, "html.parser")

shootingstats2023 = soup.find(id="all_shooting_stats")

shooting_2023 = pd.read_html(str(shootingstats2023))[0]
shooting_2023["Year"] = year

shooting_2023

shooting_2023.columns = shooting_2023.columns.droplevel(0)
shooting_2023

shooting_2023.columns

drop_cols = ['Rk','Unnamed: 9_level_1', 'Unnamed: 16_level_1', 'Unnamed: 23_level_1', 'Unnamed: 26_level_1', 'Unnamed: 29_level_1', 'Unnamed: 32_level_1']
shooting_2023 = shooting_2023.drop(drop_cols, axis = 1)
shooting_2023

shooting_2023.columns

shooting_cols = ['Player', 'Pos', 'Age', 'Tm', 'G', 'MP', 'FG%', 'Dist.', '2P_FGA%',
       '0-3_FGA%', '3-10_FGA%', '10-16_FGA%', '16-3P_FGA%', '3P_FGA%', '2P_FG%', '0-3_FG%', '3-10_FG%', '10-16_FG%',
       '16-3P_FG%', '3P_FG%', '2P_FGast', '3P_FGast', 'Dunks_%FGA', 'Dunks_#', 'Corner_%3PA', 'Corner_3P%', 
                 'Heaves_Att.', 'Heaves_#', 'Year']

shooting_2023.columns = shooting_cols
shooting_2023

with open("C:/Users/kevin/Jupyter Notebook Projects/defensive stats/2023.html", encoding='utf-8') as f:
    page = f.read()
    
soup = BeautifulSoup(page, "html.parser")

defensivestats2023 = soup.find(id="all_advanced_stats")

defensive_2023 = pd.read_html(str(defensivestats2023))[0]
defensive_2023["Year"] = year

defensive_2023

defensive_2023.columns

drop_cols = ['Rk', 'Unnamed: 19', 'Unnamed: 24']
defensive_2023 = defensive_2023.drop(drop_cols, axis = 1)
defensive_2023

Now that I have worked out the logic for getting the basic dataframe for both the shooting and defensive metrics, I can integrate the logic into a loop that adds info for each year.

### Loop for Shooting Stats

dfs = []
for year in years:
    with open("C:/Users/kevin/Jupyter Notebook Projects/shooting stats/{}.html".format(year), encoding='utf-8') as f:
        page = f.read()
    soup = BeautifulSoup(page, "html.parser")
    shootingstats = soup.find(id="all_shooting_stats")
    shooting = pd.read_html(str(shootingstats))[0]
    shooting["Year"] = year
    
    shooting.columns = shooting.columns.droplevel(0)
    
    drop_cols = ['Rk','Unnamed: 9_level_1', 'Unnamed: 16_level_1', 'Unnamed: 23_level_1', 'Unnamed: 26_level_1', 'Unnamed: 29_level_1', 'Unnamed: 32_level_1']
    shooting = shooting.drop(drop_cols, axis = 1)
    
    shooting_cols = ['Player', 'Pos', 'Age', 'Tm', 'G', 'MP', 'FG%', 'Dist.', '2P_FGA%',
       '0-3_FGA%', '3-10_FGA%', '10-16_FGA%', '16-3P_FGA%', '3P_FGA%', '2P_FG%', '0-3_FG%', '3-10_FG%', '10-16_FG%',
       '16-3P_FG%', '3P_FG%', '2P_FGast', '3P_FGast', 'Dunks_%FGA', 'Dunks_#', 'Corner_%3PA', 'Corner_3P%', 
                 'Heaves_Att.', 'Heaves_#', 'Year']

    shooting.columns = shooting_cols
    
    dfs.append(shooting)

shooting_df = pd.concat(dfs)

shooting_df.to_csv('C:/Users/kevin/Downloads/shooting stats 2014-2023.csv', index=False)

shooting_df = pd.read_csv("C:/Users/kevin/Downloads/shooting stats 2014-2023.csv")

### Loop for Defensive Data

dfs2 = []
for year in years:
    with open("C:/Users/kevin/Jupyter Notebook Projects/defensive stats/{}.html".format(year), encoding='utf-8') as f:
        page = f.read()
    soup = BeautifulSoup(page, "html.parser")
    defensivestats = soup.find(id="all_advanced_stats")
    defensive = pd.read_html(str(defensivestats))[0]
    defensive["Year"] = year
    
    drop_cols = ['Rk', 'Unnamed: 19', 'Unnamed: 24']
    defensive = defensive.drop(drop_cols, axis = 1)
    
    dfs2.append(defensive)

defensive_df = pd.concat(dfs2)

defensive_df.to_csv('C:/Users/kevin/Downloads/defensive stats 2014-23.csv', index=False)

defensive_df = pd.read_csv("C:/Users/kevin/Downloads/defensive stats 2014-2023.csv")

### Scraping Measurement Data from NBA API

import nba_api

custom_headers = {
    'Host': 'stats.nba.com',
    'Connection': 'keep-alive',
    'Cache-Control': 'max-age=0',
    'Upgrade-Insecure-Requests': '1',
    'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/73.0.3683.86 Safari/537.36',
    'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3',
    'Accept-Encoding': 'gzip, deflate, br',
    'Accept-Language': 'en-US,en;q=0.9',
}

Below is the player dictionary I have previously saved from the NBA API that will serve as keys to access other endpoints.

player_dict = pd.read_csv("C:/Users/kevin/Downloads/player dictionary.csv")

from nba_api.stats.library.parameters import SeasonAll
from nba_api.stats.library.parameters import SeasonType

from nba_api.stats.endpoints import commonplayerinfo

Finding Scottie Barnes's ID from the player dictionary as an example to pull from the endpoint.

player_dict[player_dict['full_name'] == 'Scottie Barnes']

commonplayerinfo.CommonPlayerInfo(player_id = '1630567').get_data_frames()[0]

I wanted height and weight as the main pre-draft indicators and this has that.

merged_df = pd.merge(shooting_df, defensive_df, on=['Player', 'Year'], how='outer')

merged_list = merged_df['Player'].unique()

player_bios = pd.DataFrame()

for i, name in enumerate(merged_list, start=1):
    # Check if 'full_name' column exists in player_dict
    if 'full_name' in player_dict.columns:
        # Get player_id for the current name if 'full_name' column exists
        matching_rows = player_dict.loc[player_dict['full_name'] == name, 'id']

        # Check if any matching rows were found
        if not matching_rows.empty:
            id = matching_rows.values[0]

            # Make the API call
            result_df = commonplayerinfo.CommonPlayerInfo(player_id=id).get_data_frames()[0]

            # Append result_df to the merged_data DataFrame
            player_bios = player_bios.append(result_df, ignore_index=True)

            # Print the player number
            print(f"Processed player {i}/{len(merged_list)}: {name}")
            
            # Introduce a random time lag
            lag = np.random.uniform(low=0.3, high=3)
            print(f'...waiting {round(lag, 1)} seconds')
            time.sleep(lag)
        else:
            print(f"No matching rows found for {name} in player_dict.")
    else:
        print("Column 'full_name' not found in player_dict.")

player_bios.to_csv('C:/Users/kevin/Downloads/player bios 2014-2023.csv', index=False)

player_bios = pd.read_csv('C:/Users/kevin/Downloads/player bios 2014-2023.csv')
