==> 2. NBA Skills Improvement Project Part 2 - Data Cleaning.ipynb <==
## NBA Skills Improvement Project Part 2 - Data Cleaning

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

pd.set_option('display.max_columns', None)
pd.reset_option('display.max_rows')

shooting_df = pd.read_csv("C:/Users/kevin/Downloads/shooting stats 2014-2023.csv")

defensive_df = pd.read_csv("C:/Users/kevin/Downloads/defensive stats 2014-23.csv")

player_bios = pd.read_csv('C:/Users/kevin/Downloads/player bios 2014-2023.csv')

### Cleaning the Shooting Data Set First

Normalizing the names to not include special characters.

shooting_df["Player"] = shooting_df["Player"].str.replace("*","", regex=False)
shooting_df["Player"] = shooting_df["Player"].str.replace("+","", regex=False)

shooting = pd.merge(shooting_df, player_bios, how='left', left_on='Player', right_on='DISPLAY_FIRST_LAST')

shooting = shooting[shooting['Player'] != 'Player']

Consolidating players down to one row if they played multiple seasons.

def single_year(df):
    if df["Tm"].nunique() == 1:
        return df  # Return the row if player only played for one team
    else:
        return df[df["Tm"] == "TOT"]

shooting = shooting.groupby(["Player", "Year"]).apply(single_year)

shooting.index = shooting.index.droplevel()
shooting.index = shooting.index.droplevel()

shooting[shooting['Player'] == 'Spencer Dinwiddie']

shooting[shooting['DISPLAY_FIRST_LAST'].isnull()]

A lot of names with accents did not merge from the player bios dataframe. This is still a small percentage of the total set so I will drop them, but the insights could end up being a little more American born driven than representative of the entire league.

shooting = shooting.dropna(subset=['DISPLAY_FIRST_LAST'])

Removed 449 rows where DISPLAY_FIRST_LAST was null, meaning I would have no data from the player bios dataframe on them.

shooting = shooting.dropna(subset=['POSITION', 'HEIGHT', 'WEIGHT'])

Removed 17 rows where position, height, or weight was null, meaning I would not have the important info I need from player bios.

shooting = shooting.drop(['JERSEY','TEAM_CODE', 'TEAM_CITY', 'TEAM_ABBREVIATION', 'TEAM_NAME', 'DRAFT_NUMBER', 'DRAFT_ROUND'], axis=1)

Removed the remaining columns with null values that were non-numeric. I did not need those columns for the analysis I am planning.

shooting.isna().sum().sort_values(ascending=False)

shooting.fillna(0, inplace=True)

The rest are numeric null values I replaced with 0. These were null likely because the players did not take shots from those distances. All null values have been cleaned.

shooting.to_csv('C:/Users/kevin/Downloads/shooting stats cleaned v2 2014-2023.csv', index=False)

### Cleaning the Defense Data Set

defensive_df["Player"] = defensive_df["Player"].str.replace("*","", regex=False)
defensive_df["Player"] = defensive_df["Player"].str.replace("+","", regex=False)

defensive_df[defensive_df['Player'] == 'Zion Williamson']

defensive = pd.merge(defensive_df, player_bios, how='left', left_on='Player', right_on='DISPLAY_FIRST_LAST')

defensive = defensive[defensive['Player'] != 'Player']

defensive = defensive.groupby(["Player", "Year"]).apply(single_year)

defensive.index = defensive.index.droplevel()
defensive.index = defensive.index.droplevel()

defensive = defensive.dropna(subset=['DISPLAY_FIRST_LAST'])

Removed 390 rows where no data was transferred from player bios.

defensive = defensive.dropna(subset=['POSITION', 'HEIGHT', 'WEIGHT'])

defensive = defensive.drop(['JERSEY','TEAM_CODE', 'TEAM_CITY', 'TEAM_ABBREVIATION', 'TEAM_NAME', 'DRAFT_NUMBER', 'DRAFT_ROUND'], axis=1)

Removed the columns I do not need again. 400 rows had null values for these.

defensive.isna().sum().sort_values(ascending=False)

Much less data that I have to fill in with zeroes here in this defensive data set. Realistically I could also drop these columns since these are not the columns I need for the data analysis, but since they are numeric values I might as well fill in with zeroes. It's also a very small number of rows that is affected at 20 out of 5000

defensive.fillna(0, inplace=True)

defensive.to_csv('C:/Users/kevin/Downloads/defensive stats cleaned v2 2014-2023.csv', index=False)
