==> 3. NBA Skills Improvement Project Part 3 - Data Manipulation.ipynb <==
## NBA Skills Improvement Project Part 3 - Data Manipulation

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

pd.set_option('display.max_columns', None)
pd.reset_option('display.max_rows')

shooting_df = pd.read_csv("C:/Users/kevin/Downloads/shooting stats cleaned v2 2014-2023.csv")
defensive_df = pd.read_csv("C:/Users/kevin/Downloads/defensive stats cleaned v2 2014-2023.csv")

### Preparing Data for Three Point Shooting Model

shooting_df[shooting_df['Player'] == 'Zion Williamson']

def convert_height(height_str):
    feet, inches = height_str.split('-')
    return (float(feet) * 12) + float(inches)

# Convert 'Height' column
shooting_df['HEIGHT'] = shooting_df['HEIGHT'].apply(convert_height)
defensive_df['HEIGHT'] = defensive_df['HEIGHT'].apply(convert_height)

Converted height to a float so I can include it in a model.

shooting_df[shooting_df['Player'] == 'Zion Williamson']

These quartiles will be very important for my analysis. These are quartiles by year for three point shooting percentage. Each player will have a number assigned for which quartile they fall into by year. If they improve in the quartile ranks from their first two years to the next two, they will be considered to have improved for the model. The same goes for the finishing in the paint model and the defensive model.

quartiles_by_year = shooting_df.groupby('Year')['3P_FG%'].quantile([0.25, 0.5, 0.75])

print("Quartiles for 3P_FG% broken out by year:")
print(quartiles_by_year)

Below is a function to match a player's row by year to the quartiles for that year to assign a quartile rank.

def assign_quartile(row):
    year = row['Year']
    percentile = row['3P_FG%']
    quartiles = quartiles_by_year.loc[year]
    if percentile <= quartiles[0.25]:
        return 1
    elif percentile <= quartiles[0.5]:
        return 2
    elif percentile <= quartiles[0.75]:
        return 3
    else:
        return 4

shooting_df['Quartile'] = shooting_df.apply(assign_quartile, axis=1)

seasons_played = shooting_df.groupby('Player').size()
filtered_shooting_df = shooting_df[shooting_df['Player'].isin(seasons_played[seasons_played >= 4].index)]

filtered_shooting_df = filtered_shooting_df[filtered_shooting_df['DRAFT_YEAR'] != 'Undrafted']

Included players in this model in the last 10 years and have played at least four seasons. Want enough data to see if someone improved.

filtered_shooting_df['DRAFT_YEAR'] = filtered_shooting_df['DRAFT_YEAR'].astype(int)
filtered_shooting_df = filtered_shooting_df[(filtered_shooting_df['DRAFT_YEAR'] >= 2013) & (filtered_shooting_df['DRAFT_YEAR'] <= 2022)]

The functions below will define the highest quartile a player achieved in their first two years as one column and the highest quartile a player achieved in the following two years as another column. If the number in the second column is higher, they are deemed to have improved; and No if not.

def max_quartile_first_2(group):
    # Consider only the first two rows
    group_first_2 = group.head(2)
    # Find the max quartile of the first two years
    max_quartile_first_2 = group_first_2['Quartile'].max()
    # Assign the max quartile to all rows
    group['FIRST_2'] = max_quartile_first_2
    return group


def max_quartile_next_2(group):
    # Consider only rows 3 and 4
    group_next_2 = group.iloc[2:4]
    # Find the max quartile of rows 3 and 4
    max_quartile_next_2 = group_next_2['Quartile'].max()
    # Assign the max quartile to all rows
    group['NEXT_2'] = max_quartile_next_2
    return group


def determine_improved(group):
    # Check if NEXT_2 is greater than FIRST_2
    if group['NEXT_2'].iloc[0] > group['FIRST_2'].iloc[0]:
        group['Improved'] = 'Yes'
    else:
        group['Improved'] = 'No'
    return group

# Apply the function to create 'FIRST_2'
filtered_shooting_df = filtered_shooting_df.groupby('Player').apply(max_quartile_first_2)

# Apply the function to create 'NEXT_2'
filtered_shooting_df = filtered_shooting_df.groupby('Player').apply(max_quartile_next_2)

# Apply the function to create 'Improved'
filtered_shooting_df = filtered_shooting_df.groupby('Player').apply(determine_improved)

Filtered out players whose max quartiles in the first two years is 4 because they theoretically could not improve in this model.

final_filtered_shooting_df = filtered_shooting_df[filtered_shooting_df['FIRST_2'] != 4]

Did the same pipeline but for a date range limited to the 2020 and 2021 seasons, so that I can use them as a test set. They have at least their first two seasons completed, which would be the training set but have yet to play two seasons after to realize what their Improvement result is.

test_shooting_df = pd.read_csv("C:/Users/kevin/Downloads/shooting stats cleaned v2 2014-2023.csv")
test_shooting_df['HEIGHT'] = test_shooting_df['HEIGHT'].apply(convert_height)

test_shooting_df['Quartile'] = test_shooting_df.apply(assign_quartile, axis=1)
test_shooting_df = test_shooting_df[test_shooting_df['DRAFT_YEAR'] != 'Undrafted']
test_shooting_df['DRAFT_YEAR'] = test_shooting_df['DRAFT_YEAR'].astype(int)
test_shooting_df = test_shooting_df[(test_shooting_df['DRAFT_YEAR'] >= 2020) & (test_shooting_df['DRAFT_YEAR'] <= 2021)]

# Apply the function to create 'FIRST_2'
test_shooting_df = test_shooting_df.groupby('Player').apply(max_quartile_first_2)

# Apply the function to create 'NEXT_2'
test_shooting_df = test_shooting_df.groupby('Player').apply(max_quartile_next_2)

# Apply the function to create 'Improved'
test_shooting_df = test_shooting_df.groupby('Player').apply(determine_improved)

final_test_shooting_df = test_shooting_df[test_shooting_df['FIRST_2'] != 4]

final_filtered_shooting_df.to_csv('C:/Users/kevin/Downloads/three point shooting model training set.csv', index=False)
final_test_shooting_df.to_csv('C:/Users/kevin/Downloads/three point shooting model test set.csv', index=False)

### Preparing Data for Finishing in the Paint Model

The pipeline for the finishing in the paint model is exactly the same as above for both training and test set.

def assign_quartile_finishing(row):
    year = row['Year']
    percentile = row['0-3_FG%']
    quartiles = quartiles_by_year_finishing.loc[year]
    if percentile <= quartiles[0.25]:
        return 1
    elif percentile <= quartiles[0.5]:
        return 2
    elif percentile <= quartiles[0.75]:
        return 3
    else:
        return 4

finishing_df = pd.read_csv("C:/Users/kevin/Downloads/shooting stats cleaned v2 2014-2023.csv")
finishing_df['HEIGHT'] = finishing_df['HEIGHT'].apply(convert_height)

quartiles_by_year_finishing = finishing_df.groupby('Year')['0-3_FG%'].quantile([0.25, 0.5, 0.75])

print("Quartiles for 0-3_FG% broken out by year:")
print(quartiles_by_year_finishing)

finishing_df['Quartile'] = finishing_df.apply(assign_quartile_finishing, axis=1)

seasons_played_finishing = finishing_df.groupby('Player').size()
finishing_df = finishing_df[finishing_df['Player'].isin(seasons_played_finishing[seasons_played_finishing >= 4].index)]

finishing_df = finishing_df[finishing_df['DRAFT_YEAR'] != 'Undrafted']
finishing_df['DRAFT_YEAR'] = finishing_df['DRAFT_YEAR'].astype(int)
finishing_df = finishing_df[(finishing_df['DRAFT_YEAR'] >= 2013) & (finishing_df['DRAFT_YEAR'] <= 2022)]

# Apply the function to create 'FIRST_2'
finishing_df = finishing_df.groupby('Player').apply(max_quartile_first_2)

# Apply the function to create 'NEXT_2'
finishing_df = finishing_df.groupby('Player').apply(max_quartile_next_2)

# Apply the function to create 'Improved'
finishing_df = finishing_df.groupby('Player').apply(determine_improved)

final_finishing_df = finishing_df[finishing_df['FIRST_2'] != 4]

test_finishing_df = pd.read_csv("C:/Users/kevin/Downloads/shooting stats cleaned v2 2014-2023.csv")
test_finishing_df['HEIGHT'] = test_finishing_df['HEIGHT'].apply(convert_height)

test_finishing_df['Quartile'] = test_finishing_df.apply(assign_quartile_finishing, axis=1)

test_finishing_df = test_finishing_df[test_finishing_df['DRAFT_YEAR'] != 'Undrafted']
test_finishing_df['DRAFT_YEAR'] = test_finishing_df['DRAFT_YEAR'].astype(int)
test_finishing_df = test_finishing_df[(test_finishing_df['DRAFT_YEAR'] >= 2020) & (test_finishing_df['DRAFT_YEAR'] <= 2021)]

# Apply the function to create 'FIRST_2'
test_finishing_df = test_finishing_df.groupby('Player').apply(max_quartile_first_2)

# Apply the function to create 'NEXT_2'
test_finishing_df = test_finishing_df.groupby('Player').apply(max_quartile_next_2)

# Apply the function to create 'Improved'
test_finishing_df = test_finishing_df.groupby('Player').apply(determine_improved)

final_test_finishing_df = test_finishing_df[test_finishing_df['FIRST_2'] != 4]

final_finishing_df.to_csv('C:/Users/kevin/Downloads/paint scoring model training set.csv', index=False)
final_test_finishing_df.to_csv('C:/Users/kevin/Downloads/paint scoring model test set.csv', index=False)

### Preparing Data for Defensive BPM Model

The pipeline for the defensive BPM model is the exact same as above except the columns are different so there are some slight changes.

def assign_quartile_defense(row):
    year = row['Year']
    percentile = row['DBPM']
    quartiles = quartiles_by_year_defense.loc[year]
    if percentile <= quartiles[0.25]:
        return 1
    elif percentile <= quartiles[0.5]:
        return 2
    elif percentile <= quartiles[0.75]:
        return 3
    else:
        return 4

defense_df = pd.read_csv("C:/Users/kevin/Downloads/defensive stats cleaned v2 2014-2023.csv")
defense_df['HEIGHT'] = defense_df['HEIGHT'].apply(convert_height)

quartiles_by_year_defense = defense_df.groupby('Year')['DBPM'].quantile([0.25, 0.5, 0.75])

print("Quartiles for DBPM broken out by year:")
print(quartiles_by_year_defense)

defense_df['Quartile'] = defense_df.apply(assign_quartile_defense, axis=1)

seasons_played_defense = defense_df.groupby('Player').size()
defense_df = defense_df[defense_df['Player'].isin(seasons_played_defense[seasons_played_defense >= 4].index)]

defense_df = defense_df[defense_df['DRAFT_YEAR'] != 'Undrafted']
defense_df['DRAFT_YEAR'] = defense_df['DRAFT_YEAR'].astype(int)
defense_df = defense_df[(defense_df['DRAFT_YEAR'] >= 2013) & (defense_df['DRAFT_YEAR'] <= 2022)]

# Apply the function to create 'FIRST_2'
defense_df = defense_df.groupby('Player').apply(max_quartile_first_2)

# Apply the function to create 'NEXT_2'
defense_df = defense_df.groupby('Player').apply(max_quartile_next_2)

# Apply the function to create 'Improved'
defense_df = defense_df.groupby('Player').apply(determine_improved)

final_defense_df = defense_df[defense_df['FIRST_2'] != 4]

test_defense_df = pd.read_csv("C:/Users/kevin/Downloads/defensive stats cleaned v2 2014-2023.csv")
test_defense_df['HEIGHT'] = test_defense_df['HEIGHT'].apply(convert_height)

quartiles_by_year_defense = defense_df.groupby('Year')['DBPM'].quantile([0.25, 0.5, 0.75])

test_defense_df['Quartile'] = test_defense_df.apply(assign_quartile_defense, axis=1)

test_defense_df = test_defense_df[test_defense_df['DRAFT_YEAR'] != 'Undrafted']
test_defense_df['DRAFT_YEAR'] = test_defense_df['DRAFT_YEAR'].astype(int)
test_defense_df = test_defense_df[(test_defense_df['DRAFT_YEAR'] >= 2020) & (test_defense_df['DRAFT_YEAR'] <= 2021)]

# Reset the index to avoid ambiguity
test_defense_df.reset_index(drop=True, inplace=True)

# Apply the function to create 'FIRST_2'
test_defense_df = test_defense_df.groupby('Player').apply(max_quartile_first_2)

# Apply the function to create 'NEXT_2'
test_defense_df = test_defense_df.groupby('Player').apply(max_quartile_next_2)

# Apply the function to create 'Improved'
test_defense_df = test_defense_df.groupby('Player').apply(determine_improved)

final_test_defense_df = test_defense_df[test_defense_df['FIRST_2'] != 4]

final_defense_df.to_csv('C:/Users/kevin/Downloads/defensive BPM model training set.csv', index=False)
final_test_defense_df.to_csv('C:/Users/kevin/Downloads/defensive BPM model test set.csv', index=False)

### Checking the data looks right

The data for each of these training and test sets look right because no one whose max quartile in their first 2 seasons was a 4 was included in the data sets. This means everyone we will analyze in part 4 had an opportunity to improve.

final_filtered_shooting_df['FIRST_2'].value_counts()

final_test_shooting_df['FIRST_2'].value_counts()

final_finishing_df['FIRST_2'].value_counts()

final_test_finishing_df['FIRST_2'].value_counts()

final_defense_df['FIRST_2'].value_counts()

final_test_defense_df['FIRST_2'].value_counts()
